Move widget helpers into utils to break circular dependency
SQUASH
